+++ CHANGELOG.md
@@ -20,7 +20,10 @@
 
 * issue: Monitor continues to throw errors about monitor column being null
 
+* feature#115: Location in UI in default view
+
 * feature#127: Limit the amount of devices shown
+
 
 --- 2.5 ---